WIP: MDEV-36024 regression with innodb_encrypt_log
Note: This change breaks crash recovery for innodb_encrypt_log.
Its only purpose is to estimate the performance when the inline
function encryption_crypt() is split in such a way that
encryption_ctx_update() will be invoked multiple times.

Even if this change worked as intended, it would be incompatible with
the existing innodb_encrypt_log format.