MDEV-35353: write rows_examined for union_all queries
For all union_all, as well as union queries, number of rows_examined
was tracked and updated correctly. However, they are only being
written to the slow query log in the case of UNION queries, but not
for UNION_ALL queries.

This PR addresses the concern to write rows_examined for UNION_ALL
queries as well to the slow query log.